Added msdf-atlas-gen submodule + basic example in Font.cpp

+++ Editor/Source/Panels/ContentBrowserPanel.cpp
@@ -46,7 +46,8 @@ namespace Gaze
             ImGui::PushID(filenameString.c_str());
             Ref<Texture2D> icon = directoryEntry.is_directory() ? m_DirectoryIcon : m_FileIcon;
             ImGui::PushStyleColor(ImGuiCol_Button, ImVec4(0, 0, 0, 0));
-            ImGui::ImageButton((ImTextureID)icon->GetRendererID(), {thumbnailSize, thumbnailSize}, {0, 1}, {1, 0});
+            ImGui::ImageButton(
+                (ImTextureID)(uint64_t)icon->GetRendererID(), {thumbnailSize, thumbnailSize}, {0, 1}, {1, 0});
 
             if (ImGui::BeginDragDropSource())
             {
